# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contrataciones:
    get:
      operationId: getContrataciones
      tags: [contrataciones]
      summary: Listar contrataciones con filtros
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: ubigeo
          in: query
          schema:
            type: ["string", "null"]
        - name: tipo
          in: query
          schema:
            type: ["string", "null"]
        - name: estado
          in: query
          schema:
            type: ["string", "null"]
        - name: procedimiento
          in: query
          schema:
            type: ["string", "null"]
        - name: entidadRuc
          in: query
          schema:
            type: ["string", "null"]
        - name: proveedorRuc
          in: query
          schema:
            type: ["string", "null"]
        - name: fechaDesde
          in: query
          schema:
            type: ["string", "null"]
        - name: fechaHasta
          in: query
          schema:
            type: ["string", "null"]
        - name: montoMin
          in: query
          schema:
            type: ["number", "null"]
        - name: montoMax
          in: query
          schema:
            type: ["number", "null"]
        - name: q
          in: query
          schema:
            type: ["string", "null"]
        - name: ordenar
          in: query
          schema:
            type: ["string", "null"]
            enum: ["fecha_desc", "fecha_asc", "monto_desc", "monto_asc", null]
      responses:
        "200":
          description: Lista paginada de contrataciones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContratacionesResponse"
  /contrataciones/{ocid}:
    get:
      operationId: getContratacion
      tags: [contrataciones]
      summary: Detalle de una contratación por OCID
      parameters:
        - name: ocid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalle de la contratación
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContratacionDetalle"
        "404":
          description: No encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /distritos:
    get:
      operationId: getDistritos
      tags: [distritos]
      summary: Listar distritos de Junín con resumen estadístico
      parameters:
        - name: provincia
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Lista de distritos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DistritoResumen"
  /distritos/{ubigeo}:
    get:
      operationId: getDistrito
      tags: [distritos]
      summary: Detalle de un distrito con sus estadísticas
      parameters:
        - name: ubigeo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalle del distrito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistritoDetalle"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Contratacion:
      type: object
      properties:
        ocid:
          type: string
        nomenclatura:
          type: ["string", "null"]
        titulo:
          type: string
        descripcion:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
        procedimiento:
          type: ["string", "null"]
        estado:
          type: ["string", "null"]
        entidadRuc:
          type: ["string", "null"]
        entidadNombre:
          type: ["string", "null"]
        proveedorRuc:
          type: ["string", "null"]
        proveedorNombre:
          type: ["string", "null"]
        ubigeoCodigo:
          type: ["string", "null"]
        ubigeoDistrito:
          type: ["string", "null"]
        ubigeoProvincia:
          type: ["string", "null"]
        montoReferencial:
          type: ["number", "null"]
        montoAdjudicado:
          type: ["number", "null"]
        moneda:
          type: ["string", "null"]
        fechaConvocatoria:
          type: ["string", "null"]
        fechaAdjudicacion:
          type: ["string", "null"]
        fechaContrato:
          type: ["string", "null"]
        plazoEjecucionDias:
          type: ["integer", "null"]
        createdAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
      required:
        - ocid
        - titulo
    ContratacionDetalle:
      allOf:
        - $ref: "#/components/schemas/Contratacion"
        - type: object
          properties:
            entidad:
              $ref: "#/components/schemas/EntidadBasica"
            proveedor:
              $ref: "#/components/schemas/ProveedorBasico"
            ubigeo:
              $ref: "#/components/schemas/UbigeoBasico"
            rawOcds:
              type: ["object", "null"]
          required: []
    ContratacionesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contratacion"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - data
        - total
        - page
        - limit
        - pages
    UbigeoBasico:
      type: object
      properties:
        codigo:
          type: string
        departamento:
          type: string
        provincia:
          type: string
        distrito:
          type: string
        latitud:
          type: ["number", "null"]
        longitud:
          type: ["number", "null"]
      required:
        - codigo
        - departamento
        - provincia
        - distrito
    DistritoResumen:
      type: object
      properties:
        codigo:
          type: string
        departamento:
          type: string
        provincia:
          type: string
        distrito:
          type: string
        latitud:
          type: ["number", "null"]
        longitud:
          type: ["number", "null"]
        superficieKm2:
          type: ["number", "null"]
        totalContrataciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
        contratacionesEsteAnio:
          type: integer
        montoEsteAnio:
          type: ["number", "null"]
      required:
        - codigo
        - departamento
        - provincia
        - distrito
        - totalContrataciones
        - contratacionesEsteAnio
    DistritoDetalle:
      allOf:
        - $ref: "#/components/schemas/DistritoResumen"
        - type: object
          properties:
            contratacionesRecientes:
              type: array
              items:
                $ref: "#/components/schemas/Contratacion"
            topProveedores:
              type: array
              items:
                $ref: "#/components/schemas/ProveedorRanking"
            distribucionTipo:
              type: array
              items:
                $ref: "#/components/schemas/TipoDistribucion"
            evolucionMensual:
              type: array
              items:
                $ref: "#/components/schemas/PuntoEvolucion"
          required: []
    EntidadBasica:
      type: object
      properties:
        ruc:
          type: string
        nombre:
          type: string
        tipo:
          type: ["string", "null"]
        nivelGobierno:
          type: ["string", "null"]
        ubigeoCodigo:
          type: ["string", "null"]
      required:
        - ruc
        - nombre
...
    ProveedorBasico:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        ubigeoCodigo:
          type: ["string", "null"]
        vigenteRnp:
          type: ["string", "null"]
      required:
        - ruc
        - razonSocial
...
    ProveedorRanking:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        totalAdjudicaciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - ruc
        - razonSocial
        - totalAdjudicaciones
...
    PuntoEvolucion:
      type: object
      properties:
        mes:
          type: string
        anio:
          type: integer
        periodo:
          type: string
        totalContrataciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - mes
        - anio
        - periodo
        - totalContrataciones
    TipoDistribucion:
      type: object
      properties:
        tipo:
          type: string
        total:
          type: integer
        monto:
          type: ["number", "null"]
        porcentaje:
          type: number
      required:
        - tipo
        - total
        - porcentaje
...