# lib/api-zod/src/generated/types/contratacionDetalle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
 * OpenAPI spec version: 0.1.0
 */
import type { Contratacion } from "./contratacion";
import type { ContratacionDetalleRawOcds } from "./contratacionDetalleRawOcds";
import type { EntidadBasica } from "./entidadBasica";
import type { ProveedorBasico } from "./proveedorBasico";
import type { UbigeoBasico } from "./ubigeoBasico";

export type ContratacionDetalle = Contratacion & {
  entidad?: EntidadBasica;
  proveedor?: ProveedorBasico;
  ubigeo?: UbigeoBasico;
  /** @nullable */
  rawOcds?: ContratacionDetalleRawOcds;
};