# lib/api-zod/src/generated/types/proveedorBasico.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
 * OpenAPI spec version: 0.1.0
 */

export interface ProveedorBasico {
  ruc: string;
  razonSocial: string;
  /** @nullable */
  ubigeoCodigo?: string | null;
  /** @nullable */
  vigenteRnp?: string | null;
}