# lib/api-zod/src/generated/types/tipoDistribucion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
 * OpenAPI spec version: 0.1.0
 */

export interface TipoDistribucion {
  tipo: string;
  total: number;
  /** @nullable */
  monto?: number | null;
  porcentaje: number;
}